test(loader): refactor and add test for calendar_dates gtfs
