+++ package.json
@@ -7,29 +7,29 @@
     "lint": "./node_modules/.bin/eslint lib",
     "test": "npm run test-download && npm run test-mysql && npm run test-mysql-spatial && npm run test-postgres && npm run test-postgis && npm run test-sqlite",
     "test-download": "mocha tests/download.test.js",
-    "test-mysql": "cross-env DIALECT=mysql mocha tests/db.test.js",
-    "test-mysql-spatial": "cross-env DIALECT=mysql-spatial mocha tests/db.test.js",
-    "test-postgres": "cross-env DIALECT=postgres mocha tests/db.test.js",
-    "test-postgis": "cross-env DIALECT=postgis mocha tests/db.test.js",
-    "test-sqlite": "cross-env DIALECT=sqlite mocha tests/db.test.js",
+    "test-mysql": "cross-env DIALECT=mysql mocha tests/db.*.test.js",
+    "test-mysql-spatial": "cross-env DIALECT=mysql-spatial mocha tests/db.*.test.js",
+    "test-postgres": "cross-env DIALECT=postgres mocha tests/db.*.test.js",
+    "test-postgis": "cross-env DIALECT=postgis mocha tests/db.*.test.js",
+    "test-sqlite": "cross-env DIALECT=sqlite mocha tests/db.*.test.js",
     "cover-all": "npm run cover-download && npm run cover-mysql && npm run cover-mysql-spatial && npm run cover-postgres && npm run cover-postgis && npm run cover-sqlite && npm run merge-coverage",
     "precover-download": "rimraf coverage && rimraf coverage-download",
     "cover-download": "./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha -- tests/download.test.js",
     "postcover-download": "cross-env DIALECT=download node scripts/renameCoverageFolder.js",
     "precover-mysql": "rimraf coverage && rimraf coverage-mysql",
-    "cover-mysql": "cross-env DIALECT=mysql ./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha -- tests/db.test.js",
+    "cover-mysql": "cross-env DIALECT=mysql ./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha -- tests/db.*.test.js",
     "postcover-mysql": "cross-env DIALECT=mysql node scripts/renameCoverageFolder.js",
     "precover-mysql-spatial": "rimraf coverage && rimraf coverage-mysql-spatial",
-    "cover-mysql-spatial": "cross-env DIALECT=mysql-spatial ./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha -- tests/db.test.js",
+    "cover-mysql-spatial": "cross-env DIALECT=mysql-spatial ./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha -- tests/db.*.test.js",
     "postcover-mysql-spatial": "cross-env DIALECT=mysql-spatial node scripts/renameCoverageFolder.js",
     "precover-postgres": "rimraf coverage && rimraf coverage-postgres",
-    "cover-postgres": "cross-env DIALECT=postgres ./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha -- tests/db.test.js",
+    "cover-postgres": "cross-env DIALECT=postgres ./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha -- tests/db.*.test.js",
     "postcover-postgres": "cross-env DIALECT=postgres node scripts/renameCoverageFolder.js",
     "precover-postgis": "rimraf coverage && rimraf coverage-postgis",
-    "cover-postgis": "cross-env DIALECT=postgis ./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha -- tests/db.test.js",
+    "cover-postgis": "cross-env DIALECT=postgis ./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha -- tests/db.*.test.js",
     "postcover-postgis": "cross-env DIALECT=postgis node scripts/renameCoverageFolder.js",
     "precover-sqlite": "rimraf coverage && rimraf coverage-sqlite",
-    "cover-sqlite": "cross-env DIALECT=sqlite ./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha -- tests/db.test.js",
+    "cover-sqlite": "cross-env DIALECT=sqlite ./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha -- tests/db.*.test.js",
     "postcover-sqlite": "cross-env DIALECT=sqlite node scripts/renameCoverageFolder.js",
     "premerge-coverage": "node scripts/preMerge.js",
     "merge-coverage": "./node_modules/.bin/lcov-result-merger \"coverage-*/lcov.info\" \"coverage/lcov.info\"",
