+++ tests/feed_with_only_calendar_dates/stop_times.txt
@@ -1,0 +1,5 @@
+trip_id,arrival_time,departure_time,stop_id,stop_sequence,pickup_type,drop_off_type
+trip1,12:00:00,12:00:00,LA,1,0,0
+trip1,14:00:00,14:00:00,SEA,2,0,0
+trip2,00:00:00,00:00:00,LA,1,0,0
+trip2,02:05:00,02:05:00,SEA,2,0,0
